+++ apps/studio/components/interfaces/ExplainVisualizer/utils.ts
@@ -1,0 +1,89 @@
+import {
+  Activity,
+  Database,
+  GitMerge,
+  Hash,
+  ListFilter,
+  SortAsc,
+  Zap,
+  Layers,
+  type LucideIcon,
+} from 'lucide-react'
+import type { ExplainNode } from './types'
+
+// Get human-readable description for an operation
+export function getOperationDescription(operation: string, node: ExplainNode): string {
+  const op = operation.toLowerCase()
+
+  if (op.includes('seq scan')) {
+    return 'Reads entire table row by row'
+  }
+  if (op.includes('index only scan')) {
+    return 'Reads data directly from index (fastest)'
+  }
+  if (op.includes('index scan')) {
+    return 'Uses index to find matching rows'
+  }
+  if (op.includes('bitmap index scan')) {
+    return 'Builds bitmap of matching rows from index'
+  }
+  if (op.includes('bitmap heap scan')) {
+    return 'Fetches rows using bitmap'
+  }
+  if (op.includes('hash join')) {
+    return 'Joins tables using hash lookup'
+  }
+  if (op.includes('merge join')) {
+    return 'Joins pre-sorted tables'
+  }
+  if (op.includes('nested loop')) {
+    return 'Joins by looping through each row'
+  }
+  if (op === 'hash') {
+    return 'Builds hash table for fast lookups'
+  }
+  if (op.includes('sort')) {
+    return 'Sorts rows for output or join'
+  }
+  if (op.includes('aggregate') || op.includes('group')) {
+    return 'Groups rows and calculates aggregates'
+  }
+  if (op.includes('limit')) {
+    return 'Returns only first N rows'
+  }
+  if (op.includes('materialize')) {
+    return 'Stores results in memory for reuse'
+  }
+  if (op.includes('gather')) {
+    return 'Collects results from parallel workers'
+  }
+
+  return ''
+}
+
+// Get an icon for the operation type
+export function getOperationIcon(operation: string): LucideIcon {
+  const op = operation.toLowerCase()
+  if (op === 'hash') return Hash
+  if (op.includes('hash join')) return GitMerge
+  if (op.includes('merge join')) return GitMerge
+  if (op.includes('nested loop')) return GitMerge
+  if (op.includes('join')) return Layers
+  if (op.includes('index')) return Zap
+  if (op.includes('seq scan')) return Database
+  if (op.includes('scan')) return Database
+  if (op.includes('filter')) return ListFilter
+  if (op.includes('sort')) return SortAsc
+  if (op.includes('aggregate') || op.includes('group')) return Activity
+  return Database
+}
+
+// Get a color class for the operation type
+export function getOperationColor(operation: string): string {
+  const op = operation.toLowerCase()
+  if (op.includes('seq scan')) return 'text-warning'
+  if (op.includes('index')) return 'text-brand'
+  if (op.includes('join')) return 'text-purple-500'
+  if (op.includes('sort') || op.includes('aggregate')) return 'text-blue-500'
+  return 'text-foreground'
+}
